+++ frontend/src/App/AppRoutes.js
@@ -39,6 +39,7 @@ import UnmappedFilesTableConnector from 'UnmappedFiles/UnmappedFilesTableConnect
 import getPathWithUrlBase from 'Utilities/getPathWithUrlBase';
 import UserMissingConnector from 'Wanted/UserMissing/UserMissingConnector';
 import UserUpgradesConnector from 'Wanted/UserUpgrades/UserUpgradesConnector';
+import PossibleDuplicatesPage from 'Wanted/PossibleDuplicates/PossibleDuplicatesPage';
 
 function AppRoutes(props) {
   const {
@@ -152,6 +153,11 @@ function AppRoutes(props) {
       <Route
         path="/wanted/file-upgrades"
         component={UserUpgradesConnector}
+      />
+
+      <Route
+        path="/wanted/possible-duplicates"
+        component={PossibleDuplicatesPage}
       />
 
       {/*